chore(x/core): small proto changes for overlay compatibility
Specify certain proto json tags and convert type of
QueryDataRequestsByStatusResponse.Total from uint64 to uint32 so that
its type in typescript is number instead of bigint.

+++ x/core/keeper/data_request_indexing.go
@@ -26,7 +26,7 @@ func (k Keeper) RemoveDataRequest(ctx sdk.Context, index types.DataRequestIndex,
 	return k.dataRequests.RemoveDataRequest(ctx, index, status)
 }
 
-func (k Keeper) GetDataRequestsByStatus(ctx sdk.Context, status types.DataRequestStatus, limit uint64, lastSeenIndex types.DataRequestIndex) ([]types.DataRequest, types.DataRequestIndex, uint64, error) {
+func (k Keeper) GetDataRequestsByStatus(ctx sdk.Context, status types.DataRequestStatus, limit uint32, lastSeenIndex types.DataRequestIndex) ([]types.DataRequest, types.DataRequestIndex, uint32, error) {
 	return k.dataRequests.GetDataRequestsByStatus(ctx, status, limit, lastSeenIndex)
 }
